message

+++ app/Http/Controllers/postController.php
@@ -3,14 +3,19 @@
 namespace App\Http\Controllers;
 
 use App\Post;
-//use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Auth;
+use http\Env\Response;
 use Illuminate\Http\Request;
+
+
 //use Illuminate\Support\Facades\Session;
+
 
 class postController extends Controller {
 
+
 	public function getDashboard(){
-		$posts = Post::all();
+		$posts = Post::orderBy('created_at', 'desc')->get();
 		return view('dashboard', ['posts'=>$posts]);
 	}
 
@@ -29,8 +34,26 @@ class postController extends Controller {
 	}
 
 	public function getPostDelete($post_id){
+
 		$post = Post::where('id', $post_id)->first();
+		if (Auth::user() != $post->user){
+			return redirect()->back();
+		}
 		$post->delete();
-		return redirect()->route('dashbaord');
+		return redirect()->route('dashboard')->with(['message'=>'Successfully deleted']);
+	}
+
+	public function postEditPost(Request $request){
+		$this->validate($request,[
+			'body'=>'required'
+		]);
+		$post = Post::find($request['postId']);
+		if (Auth::user() != $post->user){
+			return redirect()->back();
+		}
+
+		$post->body = $request['body'];
+		$post->update();
+		return response()->json(['new_body' => $post->body], 200);
 	}
 }
